job_match_backend/requirements.txt
asgiref==3.8.1
Django==5.1.1
gunicorn==21.2.0
sqlparse==0.4.4
dj-database-url==0.5.0
djhtml==3.0.6
pytz==2024.1
whitenoise==6.6.0
djangorestframework==3.15.1
psycopg_binary==3.1.19
psycopg2==2.9.9
whitenoise==6.6.0
djangorestframework-simplejwt==5.3.1
pyjwt==2.8.0
django-cors-headers==4.4.0
autopep8==2.3.1
pycodestyle==2.12.0
pydotplus==2.0.2
python-dotenv==1.0.1